# laravel-bek/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Patreon Clone API",
        "description": "API dokumentacija za platformu za podršku kreatorima",
        "contact": {
            "name": "API Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/creators/{id}/tiers": {
            "get": {
                "tags": [
                    "Tiers"
                ],
                "summary": "List tiers (subscription levels) of a creator",
                "description": "Display a listing of tiers for a specific creator.",
                "operationId": "6792389268aeb2dc33f2884674690653",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Creator ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tiers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TierResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Creator not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tiers"
                ],
                "summary": "Create a new tier (creator only)",
                "description": "Add a sub tier",
                "operationId": "a88f170a321544cbbc13406d369596d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Creator ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "cena_mesecno"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Gold"
                                    },
                                    "cena_mesecno": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9.99
                                    },
                                    "opis": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tier created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TierResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden (not your creator profile)"
                    },
                    "404": {
                        "description": "Creator not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tiers/{id}": {
            "put": {
                "tags": [
                    "Tiers"
                ],
                "summary": "Update a tier (creator only)",
                "description": "Update existing a sub tier",
                "operationId": "9825cc5f316019bbbcac57952afc2ec0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "naziv": {
                                        "type": "string"
                                    },
                                    "cena_mesecno": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "opis": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tier updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TierResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Tier not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tiers"
                ],
                "summary": "Delete a tier (creator only)",
                "description": "Delete existing a sub tier",
                "operationId": "fb9ff3d68522107f5ab04e9f11e82448",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tier ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tier deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Tier not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-tiers": {
            "get": {
                "tags": [
                    "Tiers"
                ],
                "summary": "Vrati nivoe kreatora sa brojem pretplatnika",
                "description": "Vraca sve pretplate ulogovanog kreatora, svaku sa projem aktivnih pretplatnika. Ukljucuje i besplatan tip pretplate.",
                "operationId": "0f47e824be7cb87be0162ecaf81eb1af",
                "responses": {
                    "200": {
                        "description": "Nivoi kreatora sa brojem pretplatnika",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tiers": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TierResource"
                                            }
                                        },
                                        "free_subscribers_count": {
                                            "description": "Number of active subscribers without a selected tier",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Korisnik nije kreator"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-tiers/earnings": {
            "get": {
                "tags": [
                    "Tiers"
                ],
                "summary": "Vrati zaradu kreatora, po nivou pretplate i ukupnu",
                "description": "Racuna i vraca ukupnu mesecnu zaradu prema pretplatnickom nivou i ukupnu mesecnu zaradu ulogovanog kreatoora.",
                "operationId": "e718bce7d34cf514424f5f4dea627c58",
                "responses": {
                    "200": {
                        "description": "Mesecna zarada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tiers": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "naziv": {
                                                        "type": "string"
                                                    },
                                                    "cena_mesecno": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "subscribers_count": {
                                                        "type": "integer"
                                                    },
                                                    "total_earnings": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total_earnings": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Korisnik nije kreator"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TierResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "naziv": {
                        "type": "string",
                        "example": "Bronzani nivo"
                    },
                    "cena_mesecno": {
                        "type": "number",
                        "format": "float",
                        "example": 4.99
                    },
                    "opis": {
                        "type": "string",
                        "example": "Osnovne pogodnosti",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}